+++ eamena/eamena/models/resource.py
@@ -54,8 +54,7 @@ class Resource(ArchesResource):
         elif self.entitytypeid == 'HERITAGE_RESOURCE_GROUP.E27':
             description_group['forms'][:0] = [
                 forms.SummaryForm.get_info(),
-
-                # forms.EvidenceAssessmentForm.get_info(),
+                forms.ArchaeologicalAssessmentForm.get_info(),
                 # forms.LocationForm.get_info(),
                 # forms.MeasurementvaluesForm.get_info(),
                 # forms.Classification1Form.get_info(),
